fix issue #15
this needs a bit of explanation: it's a problem which happens also in plain
pdb, because it reads source files as str and try to prints them on stdout
without doing any decoding/encoding. So, if stdout has a different encoding
than the source, you get bad characters.

The proper fix would be to detect the encoding from the magic comment in the
file (e.g. "# -*- encoding: utf-8 -*-"). However, there is no easy way in the
stdlib to do it, so I opted for something simpler: we try to decode the source
by trying one after the other the encodings specified in the configurable
option config.encodings, which defaults to ('utf-8', 'latin-1'). This should
work in most cases.