Fix completion after debug command
This does not restore the old completer anymore, which does not work
with nested calls to `interaction`, since the inner one (via `do_debug`)
would set the completer to `fancycompleter.Completer` again.

An alternative might be to check if it is `Pdb.complete` already (not
easy, since `fancycompleter.setup()` needs to be called and sets this
already), or use some private attribute with the new `Pdb` instance.
But I think it is good to just leave it as-is, given that it get
overwritten/set by fancycompleter anyway already.